% Auto-generated: do not edit by hand
\name{deckGL}

\alias{deckGL}

\title{DeckGL component}

\description{
This component lets you visualizes PyDeck and deck/json files directly in Dash. It also exposes various events (such as click, hover and drag) inside callbacks.
}

\usage{
deckGL(data=NULL, id=NULL, mapboxKey=NULL, clickEvent=NULL,
clickInfo=NULL, hoverEvent=NULL, hoverInfo=NULL,
dragStartEvent=NULL, dragStartInfo=NULL, dragEndEvent=NULL,
dragEndInfo=NULL)
}

\arguments{
\item{data}{Named list | character. Your map using the Deck.gl JSON format. This can be generated by calling
pdk.Deck(...).to_json(). Both a Python dictionary and a JSON-string your map is accepted.}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{mapboxKey}{Character. `mapboxKey` (text): You will need a mapbox token to use deck.gl. Please create a mapbox
and follow the instructions here: https://docs.mapbox.com/help/how-mapbox-works/access-tokens/}

\item{clickEvent}{Named list. This prop is updated when an element in the map is clicked. This contains
the original gesture event (in JSON).}

\item{clickInfo}{Named list. This prop is updated when an element in the map is clicked. This contains
the picking info describing the object being clicked.

Complete description here:
https://deck.gl/docs/developer-guide/interactivity#the-picking-info-object}

\item{hoverEvent}{Named list. This prop is updated when an element in the map is hovered. This contains
the original gesture event (in JSON).}

\item{hoverInfo}{Named list. This prop is updated when an element in the map is hovered. This contains
the picking info describing the object being hovered.

Complete description here:
https://deck.gl/docs/developer-guide/interactivity#the-picking-info-object}

\item{dragStartEvent}{Named list. This prop is updated when the user starts dragging on the canvas. This contains
the original gesture event (in JSON).}

\item{dragStartInfo}{Named list. This prop is updated when the user starts dragging on the canvas. This contains
the picking info describing the object being dragged.

Complete description here:
https://deck.gl/docs/developer-guide/interactivity#the-picking-info-object}

\item{dragEndEvent}{Named list. This prop is updated when the user releases from dragging the canvas. This contains
the original gesture event (in JSON).}

\item{dragEndInfo}{Named list. This prop is updated when the user releases from dragging the canvas. This contains
the picking info describing the object being dragged.

Complete description here:
https://deck.gl/docs/developer-guide/interactivity#the-picking-info-object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

